+++ test/connection/test_connector.py
@@ -9,139 +9,28 @@ event. These events are handled by the Reader, Writer, and Dispatcher.
 """
 
 import asyncio
-import logging
 
 import pytest
 
+from photonpump.connection import Connector, ConnectorCommand
 from photonpump.conversations import Ping
-from photonpump.connection import Connector, ConnectorCommand, MessageDispatcher
 from photonpump.discovery import (
     DiscoveryFailed, NodeService, SingleNodeDiscovery
 )
 
+from .fakes import EchoServer, TeeQueue, SpyDispatcher
+
 
 async def connector_event(connector_event):
     fut = asyncio.Future()
+
     def _cb(*args):
         if not fut.done():
             fut.set_result(None)
+
     connector_event.append(_cb)
+
     return await fut
-
-
-class TeeQueue:
-
-    def __init__(self):
-        self.items = []
-        self.queue = asyncio.Queue()
-        self.teed_queue = asyncio.Queue()
-
-    async def get(self):
-        return await self.queue.get()
-
-    async def put(self, item):
-        self.items.append(item)
-        await self.queue.put(item)
-        await self.teed_queue.put(item)
-
-    async def next_event(self, count=None):
-        if not count:
-            return await self.teed_queue.get()
-        needed = count
-        result = []
-
-        while needed > 0:
-            result.append(await self.teed_queue.get())
-            needed -= 1
-
-        return result
-
-
-class EchoServerClientProtocol(asyncio.Protocol):
-
-    def __init__(self, cb, number):
-        self.cb = cb
-        self.number = number
-
-    def connection_made(self, transport):
-        peername = transport.get_extra_info('peername')
-        logging.info('Connection from {} to protocol {}'.format(peername, self.number))
-        self.transport = transport
-        self.cb(transport)
-
-    def data_received(self, data):
-        # message = data
-        logging.info('ServerProtocol {} received data: {!r}'.format(self.number, data))
-
-        # print('Send: {!r}'.format(message))
-        self.transport.write(data)
-
-        # print('Close the client socket')
-        # self.transport.close()
-
-
-class EchoServer:
-
-    def __init__(self, addr, loop):
-        self.host = addr.address
-        self.port = addr.port
-        self.loop = loop
-        self.protocol_counter = 0
-        self.running = False
-
-    async def __aenter__(self):
-        self.transports = []
-        server = self.loop.create_server(
-            self.make_protocol, self.host, self.port
-        )
-        self._server = await server
-        self.running = True
-        logging.info("Echo server is running %s", self._server)
-
-        return self
-
-    def make_protocol(self):
-        self.protocol_counter += 1
-
-        return EchoServerClientProtocol(
-            self.transports.append, self.protocol_counter
-        )
-
-    async def __aexit__(self, exc_type, exc, tb):
-        self.stop()
-
-    def stop(self):
-        if not self.running:
-            return
-
-        for transport in self.transports:
-            transport.close()
-        self._server.close()
-        self.running = False
-
-
-class SpyDispatcher:
-
-    def __init__(self):
-        self.received = TeeQueue()
-        self.pending_messages = None
-        self.active_conversations = {}
-
-    async def dispatch(self, msg, _):
-        logging.info("Received inbound message %s", msg)
-        await self.received.put(msg)
-
-    async def start_conversation(self, conversation):
-        if self.pending_messages:
-            await self.pending_messages.put(conversation.start())
-        self.active_conversations[conversation.conversation_id] = (conversation, None)
-
-    async def write_to(self, output):
-        self.pending_messages = output
-        for (conversation, _) in self.active_conversations.values():
-            await self.pending_messages.put(conversation.start())
-
-
 
 
 @pytest.mark.asyncio
@@ -159,7 +48,9 @@ async def test_when_connecting_to_a_server(event_loop):
     async with EchoServer(addr, event_loop):
 
         dispatcher = SpyDispatcher()
-        connector = Connector(SingleNodeDiscovery(addr), dispatcher, loop=event_loop)
+        connector = Connector(
+            SingleNodeDiscovery(addr), dispatcher, loop=event_loop
+        )
 
         ping = Ping()
 
@@ -187,7 +78,10 @@ async def test_when_a_server_disconnects(event_loop):
 
     dispatcher = SpyDispatcher()
     connector = Connector(
-        SingleNodeDiscovery(addr), dispatcher, loop=event_loop, ctrl_queue=queue
+        SingleNodeDiscovery(addr),
+        dispatcher,
+        loop=event_loop,
+        ctrl_queue=queue
     )
     raised_disconnected_event = asyncio.Future(loop=event_loop)
 
@@ -213,7 +107,6 @@ async def test_when_a_server_disconnects(event_loop):
         reconnect = await queue.next_event()
         assert reconnect.command == ConnectorCommand.Connect
 
-
     assert raised_disconnected_event.result() is True
     await connector.stop()
 
@@ -230,7 +123,10 @@ async def test_when_three_heartbeats_fail_in_a_row(event_loop):
     addr = NodeService("localhost", 8338, None)
     dispatcher = SpyDispatcher()
     connector = Connector(
-        SingleNodeDiscovery(addr), dispatcher, loop=event_loop, ctrl_queue=queue
+        SingleNodeDiscovery(addr),
+        dispatcher,
+        loop=event_loop,
+        ctrl_queue=queue
     )
 
     async with EchoServer(addr, event_loop):
@@ -264,7 +160,10 @@ async def test_when_a_heartbeat_succeeds(event_loop):
     addr = NodeService("localhost", 8338, None)
     dispatcher = SpyDispatcher()
     connector = Connector(
-        SingleNodeDiscovery(addr), dispatcher, loop=event_loop, ctrl_queue=queue
+        SingleNodeDiscovery(addr),
+        dispatcher,
+        loop=event_loop,
+        ctrl_queue=queue
     )
 
     async with EchoServer(addr, event_loop):
@@ -331,9 +230,7 @@ async def test_when_discovery_fails_on_reconnection(event_loop):
     await connector.start()
     [connect, connection_failed] = await queue.next_event(count=2)
 
-    [reconnect, failed] = await asyncio.wait_for(
-        queue.next_event(count=2), 2
-    )
+    [reconnect, failed] = await asyncio.wait_for(queue.next_event(count=2), 2)
     assert failed.command == ConnectorCommand.HandleConnectorFailed
     assert policy.recorded == addr
     assert isinstance(await wait_for_stopped, DiscoveryFailed)
@@ -345,7 +242,10 @@ async def test_when_the_connection_fails_with_an_error(event_loop):
     addr = NodeService("localhost", 8338, None)
     dispatcher = SpyDispatcher()
     connector = Connector(
-        SingleNodeDiscovery(addr), dispatcher, loop=event_loop, ctrl_queue=queue
+        SingleNodeDiscovery(addr),
+        dispatcher,
+        loop=event_loop,
+        ctrl_queue=queue
     )
 
     async with EchoServer(addr, event_loop):
@@ -368,7 +268,10 @@ async def test_when_restarting_a_running_connector(event_loop):
     addr = NodeService("localhost", 8338, None)
     dispatcher = SpyDispatcher()
     connector = Connector(
-        SingleNodeDiscovery(addr), dispatcher, loop=event_loop, ctrl_queue=queue
+        SingleNodeDiscovery(addr),
+        dispatcher,
+        loop=event_loop,
+        ctrl_queue=queue
     )
 
     async with EchoServer(addr, event_loop):
@@ -392,7 +295,10 @@ async def test_when_restarting_a_stopped_connector(event_loop):
     addr = NodeService("localhost", 8338, None)
     dispatcher = SpyDispatcher()
     connector = Connector(
-        SingleNodeDiscovery(addr), dispatcher, loop=event_loop, ctrl_queue=queue
+        SingleNodeDiscovery(addr),
+        dispatcher,
+        loop=event_loop,
+        ctrl_queue=queue
     )
 
     async with EchoServer(addr, event_loop):
